pyproject.toml
[tool.poetry]
name = "asie"
version = "0.0.1"
description = ""
authors = ["Rob Nelson <nexisentertainment@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
lxml = "^4.9.2"
beautifulsoup4 = "^4.11.2"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
mypy = "^1.0.1"
black = "^23.1.0"
isort = "^5.12.0"
pybuildtools = "^0.5.5"
types-lxml = "^2023.2.11"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.isort]
profile = "black"